--- a/tests.py
+++ b/tests.py
@@ -3,7 +3,7 @@
 # from __future__ import absolute_import, division, print_function, unicode_literals
 
 from collections import OrderedDict
-from uuid import uuid4
+from uuid import uuid4, UUID
 import random
 import datetime
 import fractions
@@ -393,39 +393,48 @@ class EdnTest(unittest.TestCase):
  }
 }''', fixture, keyword_keys=True, indent=1)
 
-        fixture = (1, 2, "foo")
+        fixture = (1, Keyword("b"), "foo")
         self.check_dumps('''\
 (
   1
-  2
+  :b
   "foo"
 )''', fixture, indent=2)
 
-        fixture = [1, 2, "foo"]
+        fixture = [1, Keyword("b"), "foo"]
         self.check_dumps('''\
 [
    1
-   2
+   :b
    "foo"
 ]''', fixture, indent=3)
 
-        fixture = {1, 2, "foo"}
+        fixture = {1, 2, 3}
         self.check_dumps('''\
 #{
     1
     2
-    "foo"
-}''', fixture, indent=4)
+    3
+}''', fixture, indent=4, sort_sets=True)
 
-        fixture = {"a": 1, "b": {"c": [1, 2, 3]}, "d": {1, 2, 3}, "e": (1, 2, 3)}
+        fixture = {
+            "a": "foo",
+            "b": {"c": [Keyword("a"), Keyword("b"), Keyword("c")]},
+            "d": {1, 2, 3},
+            "e": (
+                datetime.date(2011, 10, 9),
+                UUID("urn:uuid:1e4856a2-085e-45df-93e1-41a0c7aeab2e"),
+                datetime.datetime(2012, 12, 22, 19, 40, 18, 0, tzinfo=pytz.utc)
+            )
+        }
         self.check_dumps('''\
 {
-  :a 1
+  :a "foo"
   :b {
     :c [
-      1
-      2
-      3
+      :a
+      :b
+      :c
     ]
   }
   :d #{
@@ -434,11 +443,11 @@ class EdnTest(unittest.TestCase):
     3
   }
   :e (
-    1
-    2
-    3
+    #inst "2011-10-09"
+    #uuid "1e4856a2-085e-45df-93e1-41a0c7aeab2e"
+    #inst "2012-12-22T19:40:18.000000Z"
   )
-}''', fixture, keyword_keys=True, sort_keys=True, indent=2)
+}''', fixture, keyword_keys=True, sort_keys=True, sort_sets=True, indent=2)
 
     def test_discard(self):
         for expected, edn_data in (