This adds a pretty printer similar to `json.dumps(indent=<int>)`. However, it does not follow Clojure formatting guidelines, instead formatting in a way more common to users from other languages like Python.

So it will convert this:

```clojure
{"a": 1, "b": (1, 2, 3), "c": {"d": [1, 2, 3]}, "e": {1, 2, 3}}
```

To this:

```clojure
{
  :a 1
  :b (
    1
    2
    3
  )
  :c {
    :d [
      1
      2
      3
    ]
  }
  :e #{
    1
    2
    3
  }
}
```

Instead of this:

```clojure
{
  :a 1
  :b (1 2 3)
  :c {
    :d [1 2 3]
  }
  :e #{1 2 3}
}
```

This should be already better than the current status quo (that is, no pretty printer at all).

Should fix issue https://github.com/swaroopch/edn_format/issues/39 (unless the author of the issue wanted a more Clojure-like pretty printer).

And sorry about the messy code, I didn't found a way to implement it cleaner.