@thiagokokada 2019-11-17
I discovered a problem when using strings and any other data structure that is not a map (since my use case is mostly maps I didn't find this issue until now). If someone has a good solution for this I am all ears, since the solution I found is mostly duplicating the `udump` logic.