@thiagokokada 2019-11-17
I think this PR is ready to review.