@swaroopch 2019-11-18
> > Since we have mutual recursion going on here, I'm curious if you know the performance impact?
> 
> Sorry, I don't. However I think it should be fine for small EDNs (that is my use case).
> 
> I can do some performance tests if you want.

This is an optional step. Would appreciate if you can take this additional request :)