@bfontaine 2019-11-18
I have a strange bug on [this EDN](https://gist.githubusercontent.com/bfontaine/c3463de9f1ef3a802e8fdce84bfb3adb/raw/51b99c286d040e6992d6709ce1623c7dd7a65910/bench.edn):

```pycon
>>> edn = open("bench.edn").read()
>>> from edn_format import loads, dumps
>>> dumps(loads(edn)) == edn
True
>>> dumps(loads(dumps(loads(edn)))) == edn
True
>>> dumps(loads(dumps(loads(edn), indent=2))) == edn
False
```

`dump`ing it after loading it works fine but when I add `indent=2` all I get is a vector of keys.

Without `indent=2` (same file): https://gist.githubusercontent.com/bfontaine/c3463de9f1ef3a802e8fdce84bfb3adb/raw/6214e56f4772483dd5b326b31c2ec6301816a65a/without-indent.edn
With `indent=2`: https://gist.githubusercontent.com/bfontaine/c3463de9f1ef3a802e8fdce84bfb3adb/raw/6214e56f4772483dd5b326b31c2ec6301816a65a/with-indent.edn

-----
I compared the performances of dumping this file 10 × 1000 times on both the `master` branch and this one (without `indent`): 
```
# master
         5596006 function calls (5026006 primitive calls) in 4.095 seconds
         5596006 function calls (5026006 primitive calls) in 3.922 seconds
         5596006 function calls (5026006 primitive calls) in 3.918 seconds
         5596006 function calls (5026006 primitive calls) in 4.055 seconds
         5596006 function calls (5026006 primitive calls) in 4.188 seconds
         5596006 function calls (5026006 primitive calls) in 4.303 seconds
         5596006 function calls (5026006 primitive calls) in 4.044 seconds
         5596006 function calls (5026006 primitive calls) in 4.058 seconds
         5596006 function calls (5026006 primitive calls) in 3.940 seconds
         5596006 function calls (5026006 primitive calls) in 3.997 seconds
# this branch
         6071006 function calls (5501006 primitive calls) in 4.907 seconds
         6071006 function calls (5501006 primitive calls) in 4.670 seconds
         6071006 function calls (5501006 primitive calls) in 5.659 seconds
         6071006 function calls (5501006 primitive calls) in 5.580 seconds
         6071006 function calls (5501006 primitive calls) in 5.186 seconds
         6071006 function calls (5501006 primitive calls) in 4.431 seconds
         6071006 function calls (5501006 primitive calls) in 4.289 seconds
         6071006 function calls (5501006 primitive calls) in 4.438 seconds
         6071006 function calls (5501006 primitive calls) in 4.344 seconds
         6071006 function calls (5501006 primitive calls) in 4.361 seconds
```

On another, larger file (200 dumps):
```
# master
         6601806 function calls (5506206 primitive calls) in 4.298 seconds
         6601806 function calls (5506206 primitive calls) in 4.049 seconds
         6601806 function calls (5506206 primitive calls) in 4.004 seconds
         6601806 function calls (5506206 primitive calls) in 3.977 seconds
         6601806 function calls (5506206 primitive calls) in 4.483 seconds
         6601806 function calls (5506206 primitive calls) in 4.453 seconds
         6601806 function calls (5506206 primitive calls) in 3.814 seconds
         6601806 function calls (5506206 primitive calls) in 3.768 seconds
         6601806 function calls (5506206 primitive calls) in 3.846 seconds
         6601806 function calls (5506206 primitive calls) in 5.382 seconds
# this branch
         7682806 function calls (6587206 primitive calls) in 5.802 seconds
         7682806 function calls (6587206 primitive calls) in 4.989 seconds
         7682806 function calls (6587206 primitive calls) in 4.673 seconds
         7682806 function calls (6587206 primitive calls) in 4.730 seconds
         7682806 function calls (6587206 primitive calls) in 4.650 seconds
         7682806 function calls (6587206 primitive calls) in 4.689 seconds
         7682806 function calls (6587206 primitive calls) in 4.816 seconds
         7682806 function calls (6587206 primitive calls) in 4.785 seconds
         7682806 function calls (6587206 primitive calls) in 5.059 seconds
         7682806 function calls (6587206 primitive calls) in 4.717 seconds
```

All this (small) overhead is due to `kwargs`. If you add `indent_step=0` to the arguments the overhead completely disappears:
```diff
diff --git a/edn_format/edn_dump.py b/edn_format/edn_dump.py
index cbd8669..1828ec5 100644
--- a/edn_format/edn_dump.py
+++ b/edn_format/edn_dump.py
@@ -92,7 +92,7 @@ def udump(obj,
           sort_keys=False,
           sort_sets=False,
           indent=None,
-          **kwargs):
+          indent_step=0):

     kwargs = {
         "string_encoding": string_encoding,
@@ -100,7 +100,7 @@ def udump(obj,
         "sort_keys": sort_keys,
         "sort_sets": sort_sets,
         "indent": indent,
-        "indent_step": kwargs.get("indent_step", 0)
+        "indent_step": indent_step,
     }

     if obj is None:
```