@thiagokokada 2019-11-18
> All this (small) overhead is due to kwargs. If you add indent_step=0 to the arguments the overhead completely disappears:

Well, back to a function argument I guess. Since the `_indent_objs` is now private this maybe fine, however now `indent_step` is an argument to `udump` (this may be fine though, since the actual API is `dumps` AFAIK).