@thiagokokada 2019-11-18
This is what it transformed the EDN:
```clojure
('[\n' '  :_id :index :guid :isActive :balance :picture :age :eyeColor :name '
 ':company :email :phone :address :about :registered :latitude :longitude '
 ':tags :range :friends :greeting :favoriteFruit\n'
 '  :_id :index :guid :isActive :balance :picture :age :eyeColor :name '
 ':company :email :phone :address :about :registered :latitude :longitude '
 ':tags :range :friends :greeting :favoriteFruit\n'
 '  :_id :index :guid :isActive :balance :picture :age :eyeColor :name '
 ':company :email :phone :address :about :registered :latitude :longitude '
 ':tags :range :friends :greeting :favoriteFruit\n'
 '  :_id :index :guid :isActive :balance :picture :age :eyeColor :name '
 ':company :email :phone :address :about :registered :latitude :longitude '
 ':tags :range :friends :greeting :favoriteFruit\n'
 '  :_id :index :guid :isActive :balance :picture :age :eyeColor :name '
 ':company :email :phone :address :about :registered :latitude :longitude '
 ':tags :range :friends :greeting :favoriteFruit\n'
 '  :_id :index :guid :isActive :balance :picture :age :eyeColor :name '
 ':company :email :phone :address :about :registered :latitude :longitude '
 ':tags :range :friends :greeting :favoriteFruit\n'
 ']')
```

So instead of parsing the array with maps it simply get the keys instead. I will try to investigate what is happening here.